path refactor
