+++ babel.config.js
@@ -6,7 +6,7 @@ module.exports = {
       {
         alias: {
           src: './src',
-          _jest: './_jest',
+          _jest: './_jest_',
         },
       },
     ],
