+++ jest.config.js
@@ -1,10 +1,10 @@
 module.exports = {
   preset: 'react-native',
   setupFilesAfterEnv: [
-    '<rootDir>/_jest/setup.ts',
-    '<rootDir>/_jest/setupTests.ts',
+    '<rootDir>/_jest_/setup.ts',
+    '<rootDir>/_jest_/setupTests.ts',
   ],
   transformIgnorePatterns: [
-    'node_modules/(?!(_jest-)?@?react-native|@react-native-community|@react-navigation)',
+    'node_modules/(?!(_jest_-)?@?react-native|@react-native-community|@react-navigation)',
   ],
 };
