+++ tsconfig.json
@@ -4,7 +4,7 @@
     "baseUrl": ".",
     "paths": {
         "src/*": ["./src/*"],
-        "_jest/*": ["./_jest/*"]
+        "_jest_/*": ["./_jest_/*"]
     }
   },
 }
